requirements.txt
python-telegram-bot[job-queue]
yt-dlp
Pillow
mutagen
python-dotenv
requests
python-telegram-bot